Create SECURITY.md
Click jacking, also known as a "UI redress attack", is when an attacker uses multiple transparent or opaque layers to trick a user into clicking on a button or link on another page when they were intending to click on the the top level page. Thus, the attacker is "hijacking" clicks meant for their page and routing them to another page, most likely owned by another application, domain, or both.

Using a similar technique, keystrokes can also be hijacked. With a carefully crafted combination of stylesheets, iframes, and text boxes, a user can be led to believe they are typing in the password to their email or bank account, but are instead typing into an invisible frame controlled by the attacker.

POC:

<html>
<head>UI REDRESSING</head>
<body>
<h1>WEBSITE IS VULNERABLE TO UI REDRESSING</h1>
<iframe width=100% height=80% src="https://alerts.vulmon.com/login "></iframe>
</body>
<html>